Add configurable ignore list for building and research
Some things in the game are not worth building, for instance the Tachyon
Theory science which is currently useless. At other times it may be
necessary to temporarily ignore the building of something. Users should
be able to add the names of anything they wish to ignore to a central
hash to disable automatic construction of that one thing.